Add recipe of libbpf-iotg_0.7.0.bb
Issue: LINUXEXEC-33741

This recipe and out-of-tree patch are copied from recipe of
libbpf_0.7.0.bb in https://github.com/intel/iotg-yocto-ese-main.git
with a little modification.

This recipe will be used by upcoming recipe of iotg-tsn-ref-sw.

Due to compatibility issue between iotg-tsn-ref-sw and libbpf,
the libbpf_1.4.0 in meta-openembedded layer cannot be used.

Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>